feat: move interface in page rather than factory
